Modify the attendance page to use a form vs. JS
This change modifies the attendance submission page to use form-based
submission instead of JS bindings.

In order to do this, there is now a binding between each of the
.student-icon buttons and an underlying select box that contains each of
the Attendance::STATUSES. Whenever you click on the button, there is
some JS that cycles through the options in the select box and THOSE
values are what are sent instead of using the CSS class and building
JSON to send back to the server.

The reasoning behind this change is that it uses more of the Rails
built-in functionality (e.g. HTML forms) as opposed to home-grown
JS that is harder to debug.